fixed warning from g++-7

+++ server/context/spi_dev.cpp
@@ -93,7 +93,7 @@ int SpiDev::set_word_length(uint8_t word_length_)
     return 0;
 }
 
-int SpiDev::recv(uint8_t *buffer, size_t n_bytes)
+int SpiDev::recv(uint8_t *buffer, int64_t n_bytes)
 {
     if (! is_ok())
         return -1;
