+++ server/context/spi_dev.hpp
@@ -54,7 +54,7 @@ class SpiDev
         return -1;
     }
 
-    int recv(uint8_t *buffer, size_t n_bytes);
+    int recv(uint8_t *buffer, int64_t n_bytes);
     int transfer(uint8_t *tx_buff, uint8_t *rx_buff, size_t len);
 
     template<typename T>
